Home: refactor logic
created a state for home to store notes components rendered by _createNotes, which relies on a useEffect to detect if 'notes' managed state is changed. there is a failsafe login noteApiSlice for fetchnotes() so that a loop is not created check noteApiSlice for more details.

useEffect does not loop with this setup and creation and rerendering of component is managed by _createNotes since notes will be recreated when mongodb collection is changed. there is probably am ore efficient way to do this but it works for now.